+++ src/main/java/com/example/hospitalManagement/entity/Departments.java
@@ -12,7 +12,7 @@ public class Departments {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private long id;
     @Column(name = "name")
     private String name;
     @Column(name = "description")
@@ -29,11 +29,11 @@ public class Departments {
     @OneToMany(mappedBy = "department")
     private List<Doctor> doctors;
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
