+++ src/main/java/com/example/hospitalManagement/entity/Role.java
@@ -10,7 +10,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id")
-    private Long id;
+    private long id;
     @Column(name = "name")
     private  String name;
     @Column(name = "description")
@@ -19,11 +19,11 @@ public class Role {
     @OneToMany(mappedBy = "role")
     private List<UserRoles> userRoles;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
